Use a signed integer to represent altitudes
This is to accomodate negative altitudes, which can occur for example on the ground in the Netherlands.
Migrating from u16 to i16 does decrease the maximum representable altitude, but still leaves it at more than 107,000ft, which is well above what might reasonably be encountered (eg, objective altitude for the Perlan Project is 90,000ft).

Closes #4